BACKLOG-5260 BACKLOG-4725: Don't try to remove external group when deleting a site
git-svn-id: svn://localhost/jahia/trunk@25955 69c369ed-ed09-0410-a4a1-e19a46c23626

+++ core/src/main/java/org/jahia/services/sites/JahiaSitesService.java
@@ -596,7 +596,10 @@ public class JahiaSitesService extends JahiaService {
                 @Override
                 public Boolean doInJCR(JCRSessionWrapper session) throws RepositoryException {
                     for (String groupPath : groupService.getGroupList(siteKey)) {
-                        groupService.deleteGroup(groupPath, session);
+                        if (StringUtils.startsWith(groupPath, site.getJCRLocalPath()  + "/groups/")
+                                && !StringUtils.startsWith(groupPath, site.getJCRLocalPath() + "/groups/providers/")) {
+                            groupService.deleteGroup(groupPath, session);
+                        }
                     }
                     session.save();
                     return true;
